docurefine/requirements.txt
PyPDF2
pytesseract
layoutparser
transformers
scikit-image
opencv-python
celery
redis
python-dotenv
matplotlib
flask
flower
pillow
docx2pdf
html2pdf
flask-login
flask-sqlalchemy
Flask-Limiter
sentry-sdk[flask]
reportlab
Werkzeug
bleach
prometheus_client
prometheus_flask_exporter
pdfkit